requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.18
aiosqlite==0.20.0
greenlet==3.1.1
sqlalchemy==2.0.36
presidio-analyzer==2.2.358
presidio-anonymizer==2.2.358
spacy==3.8.4
python-docx==1.1.2
openpyxl==3.1.5
odfpy==1.4.1
pdfplumber==0.11.4
PyJWT==2.10.1
cryptography==44.0.0
httpx==0.28.1
structlog==24.4.0
python-ulid==3.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
